% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_stats.R
\name{account_stats}
\alias{account_stats}
\title{account_stats}
\usage{
account_stats(
  coord_graph,
  result,
  weight_threshold = c("full", "fast", "none")
)
}
\arguments{
\item{coord_graph}{an \code{igraph} object generated by \link{generate_coordinated_network}}

\item{result}{a table generated by \link{detect_groups}}

\item{weight_threshold}{The threshold to be used for filtering the graph
(options: "full", "fast", or "none").}
}
\value{
a data.table with summary statistics for each account
}
\description{
Calculate account statistics: total posts shared, average time delta,
average edge symmetry score.
}
\details{
With this helper function, you can obtain summary statistics for the accounts in
the network. When applied to a network for which a narrower \code{time_window} has
been calculated using the \link{flag_speed_share} function, the summary statistics
are computed separately for both the full and faster networks. It is possible
to further refine the analysis by considering only the nodes that exceed the
edge weight threshold. This filter can be applied to the full network when no
faster network is identified by the \link{flag_speed_share} function,
or when the network to be filtered is the full network. Alternatively, it can
be applied to the fast network when this is preliminary identified via the
\link{flag_speed_share} function.

The share count is performed thanks to the table resulting from the
\link{detect_groups} function. If the user has used the optional
\link{flag_speed_share} function and decides to calculate the statistics on
the fastest graph (by indicating weight_threshold = "fast"), the share count
is calculated considering only the shares made in the fastest time window.
Alternatively, the shares in the largest time window are considered (option
weight_threshold = "full" or weight_threshold = "none"). When calculating the
share count, all shares made by the accounts are considered, regardless of
whether they are shares of posts resulting to be shared in a coordinated
fashion or not according to the edge weight threshold. In other words, this
is a measure of an account's activity in the time window under consideration.
}
