% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling_filter.R
\name{data_wrangling_filter}
\alias{data_wrangling_filter}
\title{data_wrangling_filter}
\usage{
data_wrangling_filter(dset_list, min_repetition)
}
\arguments{
\item{dset_list}{the list returned by \link[CooRTweet]{data_wrangling}.}

\item{min_repetition}{minimum number of repetition set by user.}
}
\value{
a pre-filtered dataframe of tweets ready for coordination analysis.
}
\description{
Filter data based on minimum thresholds to avoid unnecessary computation.
}
