% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_coordinated_groups.R
\name{do_detect_coordinated_groups}
\alias{do_detect_coordinated_groups}
\title{do_detect_coordinated_groups}
\usage{
do_detect_coordinated_groups(x, time_window = 10, min_repetition = 2)
}
\arguments{
\item{x}{a data.table with the columns: \code{object_id} (uniquely identifies
coordinated content), \code{id_user} (unique ids for users), \code{content_id}
(id of user generated content), \code{timestamp_share} (integer)}

\item{time_window}{the number of seconds within which shared contents
are to be considered as coordinated (default to 10 seconds).}

\item{min_repetition}{the minimum number of published coordinated
contents necessary for a user to be included it in the coordinated
network. (defaults to 2)}

\item{result}{The result of the previous filtering steps.}
}
\value{
a data.table with ids of coordinated contents. Columns:
\code{object_id}, \code{id_user}, \code{id_user_y}, \code{content_id}, \code{content_id_y},
\code{timedelta}

The result with loops removed.
}
\description{
Private function that actually performs \link{detect_coordinated_groups}.
}
