% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_groups.R
\name{do_remove_loops}
\alias{do_remove_loops}
\title{Remove loops from the result.}
\usage{
do_remove_loops(result)
}
\arguments{
\item{result}{The result of the previous filtering steps.}
}
\value{
The result with loops removed.
}
\description{
This function is a private utility function that removes loops (i.e., accounts
sharing their own content) from the result.
}
