% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_twitter.R
\name{dt_unnest_wider}
\alias{dt_unnest_wider}
\title{dt_unnest_wider}
\usage{
dt_unnest_wider(dt, columns, keep = FALSE, simplify = TRUE)
}
\arguments{
\item{dt}{A data.table as input}

\item{columns}{the name(s) of the column(s) to be unnested (as string)}

\item{keep}{Defaults to FALSE. Keep the column to be unnested.}

\item{simplify}{Defaults to TRUE. Try to simplify resulting columns}
}
\description{
Private utility function for \code{data.table}.
Unnest a column into wide format. Takes the first 1000 rows as reference!
Tries to emulate the functionality of tidyr::unnest_wider()
Current implementation is probably very slow.
}
