% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_weight_contribution.R
\name{edge_weight_contribution}
\alias{edge_weight_contribution}
\title{edge_weight_contribution
This function is a private utility function that calculate the contribution of each node on their
edge, and a contribution index that summarizes the equality of contributions between nodes on an
edge. When the faster network is provided, the equilibrium index is computed on the subset of
faster nodes. Otherwise, it is computed on the full network.}
\usage{
edge_weight_contribution(coord_graph, incidence_matrix, fast_net, fast_net_col)
}
\arguments{
\item{coord_graph}{The coord_graph resulting from the previous steps.}

\item{incidence_matrix}{the incidence_matrix resulting from the previous steps.}
}
\value{
The coord_graph with additional attributes from_contrib, to_contrib, and contrib_index
}
\description{
When one user share the same object_id in the defined time window many times
(such as a spammer), along with other less active users, the edge weight between them is
inflated by the hyper-activity of that user. This situation can occur in the case of (1)
outliers in the network such as a spammer, but also (2) in the case where object_ids are
prone to repetition by their very nature, such as a hashtag. The second possibility appears
rarer if object_ids are analyzed that, in a normal common context normally do not repeat
identically repeatedly, such as URLs, leaving mainly the problem of spammers.
To account for these possibilities, we compute the relative contribution of node on their
edge, and a index indicating the equality of contributions between nodes on an edge, that has
a value of 1 when the contributions are equal and 0 when there is maximum inequality.
}
