% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_weight_equilibrium_index.R
\name{edge_weight_equilibrium_index}
\alias{edge_weight_equilibrium_index}
\title{edge_weight_equilibrium_index
Calculate a edge weight equilibrium index that summarizes the contribution of each vertex
to their total edge weight. When the faster network is provided, the equilibrium index
is computed on the subset of faster nodes. Otherwise, it is computed on the full network.}
\usage{
edge_weight_equilibrium_index(g = coord_graph, x = x, df = df)
}
\arguments{
\item{g}{The coord_graph resulting from the previous steps.}

\item{x}{A data.table (result from \code{detect_coordinated_groups})}
}
\value{
The coord_graph with additional attribute edge_eq.
}
\description{
When one user share the same object_id in the defined time window many times
(such as a spammer), along with other less active users, the edge weight between them is
inflated by the hyper-activity of that user. This situation can occur in the case of (1)
outliers in the network such as a spammer, but also (2) in the case where object_ids are
prone to repetition by their very nature, such as a hashtag. The second possibility appears
rarer if object_ids are analyzed that, in a normal common context normally do not repeat
identically repeatedly, such as URLs, leaving mainly the problem of spammers.
To account for these possibilities, we define a measure of relative contribution to the
edge weight (edge_eq), which can be used to analyze or filter the resulting graph.
}
\details{
This function is a private utility function that calculates summary measures of edge
weight contribution.
}
