% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages_get_coordinated_tweets.R
\name{error_messages_get_coordinated_tweets}
\alias{error_messages_get_coordinated_tweets}
\title{error_messages_get_coordinated_tweets}
\usage{
error_messages_get_coordinated_tweets(
  data_path = data_path,
  coord_function = coord_function,
  reply_type = reply_type,
  time_window = time_window,
  min_repetition = min_repetition
)
}
\arguments{
\item{data_path}{see \link[CooRTweet]{get_coordinated_tweets}.}

\item{coord_function}{see \link[CooRTweet]{get_coordinated_tweets}.}

\item{reply_type}{see \link[CooRTweet]{get_coordinated_tweets}.}

\item{time_window}{see \link[CooRTweet]{get_coordinated_tweets}.}

\item{min_repetition}{see \link[CooRTweet]{get_coordinated_tweets}.}
}
\value{
a dataframe of tweets ready for coordination analysis.
}
\description{
Utility function with error messages for get_coordinated_tweets.
}
