% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_groups.R
\name{filter_min_participation}
\alias{filter_min_participation}
\title{Filter the result by minimum participation}
\usage{
filter_min_participation(x, result, min_participation)
}
\arguments{
\item{x}{The original data table where a preliminary filter is applied}

\item{result}{A data table containing the result data from calc_group_combinations.}

\item{min_participation}{The minimum activity threshold. accounts with participation count
greater than this threshold will be retained in the final 'result' table.}
}
\value{
A data table with filtered rows based on the specified minimum participation.
}
\description{
This private function filters the result by the minimum number of participation required.
}
