% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_coordinated_groups.R
\name{filter_min_repetition}
\alias{filter_min_repetition}
\title{Filter the result by minimum repetition.}
\usage{
filter_min_repetition(x, result, min_participation)
}
\arguments{
\item{x}{A data table from a coordination detection function}

\item{result}{A data table containing the result data.}

\item{min_participation}{The minimum repetition threshold. Users with repetition count
greater than this threshold will be retained.}
}
\value{
A data table with filtered rows based on the specified minimum repetition.
}
\description{
This private function filters the result by the minimum number of repetitions required.
}
