% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_speed_share.R
\name{flag_speed_share}
\alias{flag_speed_share}
\title{flag_speed_share}
\usage{
flag_speed_share(x, result, min_participation, time_window)
}
\arguments{
\item{x}{A data table from a coordination detection function}

\item{result}{A data table containing the result data.}

\item{min_participation}{The minimum repetition threshold. Users with repetition count
greater than this threshold will be retained (default parameter equal to
the one used in the detect_coordinated_groups function).}

\item{time_window}{The number of seconds within which shared contents are to be considered as
coordinated according to the new time_window (default parameter equal to
the one used in the detect_coordinated_groups function).}
}
\value{
A results data table that includes an additional column set to 1 when the share
corresponds with the new time_window, and 0 otherwise.
}
\description{
Function to update result based on a narrower time_window.
}
\details{
This function identifies and marks the subset of results that match a more stringent time window.
}
