% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{german_elections}
\alias{german_elections}
\title{German 2021 election campaign}
\format{
\subsection{\code{mmmp}}{

A data.frame with 218,971 rows and 7 columns:
\describe{
\item{account_id}{character, with shorthand for platform}
\item{post_id}{integer}
\item{url_id}{integer, anonymized url contained in post}
\item{hashtag_id}{integer, anonymized hashtag contained in post}
\item{domain_id}{integer, anonymized domain of url}
\item{phash_id}{integer, anonymized perceptual hash of shared image}
\item{timestamp}{numeric, timestamp of post}
}
}
}
\source{
Righetti, N., Giglietto, F., Kakavand, A. E.,
Kulichkina, A., Marino, G., & Terenzi, M. (2022).
Political Advertisement and Coordinated Behavior
on Social Media in the Lead-Up to the 2021 German Federal
Elections. Düsseldorf: Media Authority of North Rhine-Westphalia.
}
\usage{
german_elections
}
\description{
An anonymized multi-platform and multi-modal dataset
of social media messages from the 2021 German election campaign.
Includes Facebook and Twitter posts.
}
\keyword{datasets}
