% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chart.R
\name{get_chart}
\alias{get_chart}
\title{get_chart}
\usage{
get_chart(coord_graph = coord_graph)
}
\arguments{
\item{coord_graph}{the igraph object representing the coordinated network.}
}
\value{
a plot in the 'Plots' window.
}
\description{
Create the coordinated network chart.
}
