% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_stats.r
\name{group_stats}
\alias{group_stats}
\title{group_stats}
\usage{
group_stats(coord_graph, weight_threshold = c("full", "fast", "none"))
}
\arguments{
\item{coord_graph}{A result \code{igraph} generated by \link{generate_coordinated_network}}

\item{weight_threshold}{The level of the network for which to calculate the statistic.
It can be "full," "fast," or "none." The first two options are applicable only
if the data includes information on a faster network, as calculated with the
\link{flag_speed_share} function. These options preliminarily filter the nodes
based on their inclusion in the subgraph filtered by edge weight threshold
("full"), filtered by edges created in the faster time window and surpassing
the edge weight threshold in that network ("fast"), or apply to the unfiltered
graph ("none").}
}
\value{
a \code{data.table} with summary statistics
}
\description{
With this helper function, you can obtain summary statistics for the objects
in the network.
}
