% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tweets.R
\name{load_tweets}
\alias{load_tweets}
\title{load_tweets}
\usage{
load_tweets(data_path = NULL)
}
\arguments{
\item{data_path}{the path to the folder containing the JSON files returned by the function \link[academictwitteR]{get_all_tweets}.}
}
\value{
a dataframe of tweets to be used in the subsequent preprocessing procedure by data_wrangling.
}
\description{
Utility function to load a dataset retrieved from Twitter API v2. Wrapper of the function \link[academictwitteR]{bind_tweets}
}
