% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netchart.R
\name{netchart}
\alias{netchart}
\title{netchart}
\usage{
netchart(coord_graph = coord_graph)
}
\arguments{
\item{coord_graph}{the igraph object representing the coordinated network.}
}
\value{
a plot in the 'Plots' window.
}
\description{
Create the coordinated network chart.
}
