% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_tweets.r
\name{normalize_text}
\alias{normalize_text}
\title{Normalize text}
\usage{
normalize_text(x)
}
\arguments{
\item{x}{The text to be normalized.}
}
\value{
The normalized text.
}
\description{
Utility function that normalizes text by removing mentions of other users, removing "RT",
converting to lower case, and trimming whitespace.
}
