% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_twitter.R
\name{preprocess_twitter_users}
\alias{preprocess_twitter_users}
\title{preprocess_twitter_users}
\usage{
preprocess_twitter_users(users)
}
\arguments{
\item{users}{a data.table with unformatted (nested user data).}
}
\value{
a data.table with reformatted user data.
}
\description{
Reformat nested twitter user data (retrieved from Twitter v2 API).
Spreads out columns and reformats nested \code{data.table} to long format.
}
\details{
Take the Twitter user data that you loaded with
\link{load_twitter_users_json} and unnests the
following columns: \code{public_metrics} and \code{entities}.
}
