% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_tweets.r
\name{remove_hashtags}
\alias{remove_hashtags}
\title{Remove hashtags}
\usage{
remove_hashtags(x)
}
\arguments{
\item{x}{The text to be processed.}
}
\value{
The text without hashtags.
}
\description{
Utility function that removes hashtags from tags.
}
