% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{russian_coord_tweets}
\alias{russian_coord_tweets}
\title{Pro-Government Russian Tweet Dataset}
\format{
\subsection{\code{russian_coord_tweets}}{

A data frame with 35,125 rows and 4 columns:
\describe{
\item{object_id}{ID of retweeted content.
Twitter API calls this "referenced_tweet_id".}
\item{account_id}{ID of the user who tweeted. Twitter API: "author_id"}
\item{content_id}{Tweet ID.}
\item{timestamp_share}{Ingeger. Timestamp (posix time)}
}
}
}
\source{
Kulichkina, A., Righetti, N., & Waldherr, A. (2024).
Protest and repression on social media: Pro-Navalny and
pro-government mobilization dynamics and coordination
patterns on Russian Twitter. New Media & Society.
https://doi.org/10.1177/14614448241254126
}
\usage{
russian_coord_tweets
}
\description{
A anonymized dataset of Tweets.
All IDs have been obscured using sha256 algorithm.
}
\keyword{datasets}
