% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_time_window.R
\name{update_time_window}
\alias{update_time_window}
\title{update_time_window}
\usage{
update_time_window(x, result, min_repetition, time_window)
}
\arguments{
\item{x}{A data table from a coordination detection function}

\item{result}{A data table containing the result data.}

\item{min_repetition}{The minimum repetition threshold. Users with repetition count
greater than this threshold will be retained (default parameter equal to
the one used in the detect_coordinated_groups function).}

\item{time_window}{The number of seconds within which shared contents are to be considered as
coordinated according to the new time_window (default parameter equal to
the one used in the detect_coordinated_groups function).}
}
\value{
A results data table that includes an additional column set to 1 when the share
corresponds with the new time_window, and 0 otherwise.
}
\description{
Function to update result based on a new time_window.
}
\details{
This function identifies and marks the subset of results that match a more stringent time window.
}
