% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_stats.R
\name{user_stats}
\alias{user_stats}
\title{user_stats}
\usage{
user_stats(x)
}
\arguments{
\item{x}{a result data.table generated by \code{detect_coordinated_groups}}
}
\value{
a data.table with summary statistics for each user
}
\description{
Calculate user statistics: total posts shared, average time delta.
}
\details{
With this helper function you get a summary of the users in the network.
High number of posts shared and low average time delta are indicators for
highly coordinated (potentially automated) user behaviour.
The function can also be applied to a network for which a narrower time_window
has been calculated with the flag_speed_shares function. To calculate the statistics
of the users included in this fastest network, just filter the table by the relevant
column time_window_(number of seconds of fastest shares).
}
