% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_groups.R
\name{filter_min_participation}
\alias{filter_min_participation}
\title{Filter the result by minimum participation}
\usage{
filter_min_participation(x, result, min_participation)
}
\arguments{
\item{x}{A data table from a coordination detection function}

\item{result}{A data table containing the result data.}

\item{min_participation}{The minimum repetition threshold. accounts with repetition count
greater than this threshold will be retained.}
}
\value{
A data table with filtered rows based on the specified minimum participation.
}
\description{
This private function filters the result by the minimum number of participation required.
}
