% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_coordinated_network.R
\name{generate_coordinated_network}
\alias{generate_coordinated_network}
\title{generate_coordinated_network}
\usage{
generate_coordinated_network(
  x,
  fast_net = FALSE,
  edge_weight = 0.5,
  subgraph = 0
)
}
\arguments{
\item{x}{a data.table (result from \code{detect_coordinated_groups}) with the Columns: \code{object_id}, \code{id_user}, \code{id_user_y}, \code{content_id}, \code{content_id_y}, \code{timedelta}}

\item{fast_net}{If the data.table x has been updated with the restrict_time_window function and this parameter is set to TRUE, two columns weight_full and weight_fast are created, the first containing the edge weights of the full graph, the second those of the subgraph that includes the shares made in the narrower time window.}

\item{edge_weight}{The edges with weight that exceeds a threshold are marked with 0 (not exceeding) or 1 (exceeding). The threshold is expressed in percentiles of the edge weight distribution in the full network and in the faster network, and any numeric value between 0 and 1 can be assigned. The default value is "0.5" which represents the median value of the edges in the network.}

\item{subgraph}{If 1 reduces the graph to the subgraph whose edges have a value that exceeds the threshold given in the edge_weight parameter (weighted subgraph).
If 2 reduces the subgraph whose nodes exhibit coordinated behavior in the narrowest time window, as established with the restrict_time_window function, to the subgraph whose edges have a value that exceeds the threshold given in the edge_weight parameter (fast weighted subgraph).
If 3 reduces the graph to the subgraph whose nodes exhibit coordinated behavior in the narrowest time window established with the restrict_time_window function (fast subgraph).
The default value is 0, meaning that no subgraph is created.}
}
\value{
A weighted, undirected network (igraph object) where the vertices (nodes) are users and edges (links) are the membership in coordinated groups (\code{object_id})
}
\description{
This function performs the second step in coordinated detection analysis by identifying
users who repeatedly engage in identical actions within a predefined time window.
It takes the results of \code{detect_coordinated_groups} and generate a network from the data.
The function offers multiple options to identify various types of networks, allowing for
filtering based on different edge weights and facilitating the extraction of distinct subgraphs.
}
