% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_network.R
\name{generate_network}
\alias{generate_network}
\title{generate_network}
\usage{
generate_network(
  x,
  intent = c("users", "content", "objects"),
  restrict_time_window = FALSE,
  edge_weight = "median_weight",
  weighted_subgraph = FALSE,
  fast_subgraph = FALSE
)
}
\arguments{
\item{x}{a data.table (result from \code{detect_coordinated_groups}) with the Columns: \code{object_id}, \code{id_user}, \code{id_user_y}, \code{content_id}, \code{content_id_y}, \code{timedelta}}

\item{restrict_time_window}{if the data.table x has been updated with the restrict_time_window function and this parameter is set to TRUE, a dichotomous restricted_time_window attribute is added to the vertices of the graph with value 1, if the vertex does not participate in the coordinated activity in the narrower time window, and 0, otherwise (default FALSE).}

\item{edge_weight}{allows edges whose weight exceeds a certain threshold to be marked with a dichotomous 0/1 attribute. Any numeric value can be assigned. The default value is "median_weight" which represents the median value of the edges in the network.}

\item{weighted_subgraph}{if TRUE reduces the graph to the subgraph whose edges have a value that exceeds the threshold given in the edge_weight parameter (default FALSE).}

\item{fast_subgraph}{if TRUE reduces the graph to the subgraph whose nodes exhibit coordinated behavior in the narrowest time window established with the restrict_time_window function (default FALSE)}
}
\value{
A weighted, undirected network (igraph object) where the vertices (nodes) are users (or \code{content_ids}) and edges (links) are the membership in coordinated groups (\code{object_id})
}
\description{
Take the results of coordinated content detection and generate a network from the data. This function generates a two-mode (bipartite) incidence matrix first, and then projects the matrix to a weighted adjacency matrix.
}
