% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_stats.r
\name{group_stats}
\alias{group_stats}
\title{group_stats}
\usage{
group_stats(coord_graph, network = c("full", "fast"))
}
\arguments{
\item{coord_graph}{A result \code{igraph} generated by \link{generate_coordinated_network}}

\item{network}{The level of the network for which to calculate the statistic.
It can be either "full" or "fast". The latter is applicable only if the data
includes information on a faster network, as calculated with the \link{flag_speed_share}
function.}
}
\value{
a \code{data.table} with summary statistics
}
\description{
Calculate the list of objects shared in a coordinated manner and the total
number of unique accounts sharing them.
}
