% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{russian_coord_tweets}
\alias{russian_coord_tweets}
\title{Pro-Government Russian Tweet Dataset}
\format{
\subsection{\code{russian_coord_tweets}}{

A data frame with 35,125 rows and 4 columns:
\describe{
\item{object_id}{ID of retweeted content.
Twitter API calls this "referenced_tweet_id".}
\item{id_user}{ID of the user who tweeted. Twitter API: "author_id"}
\item{content_id}{Tweet ID.}
\item{timestamp_share}{Ingeger. Timestamp (posix time)}
}
}
}
\source{
Kulichkina (in Press).
}
\usage{
russian_coord_tweets
}
\description{
A anonymized dataset of Tweets.
All IDs have been obscured using sha256 algorithm.
}
\keyword{datasets}
