% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_stats.R
\name{user_stats}
\alias{user_stats}
\title{user_stats}
\usage{
user_stats(x)
}
\arguments{
\item{x}{a result data.table generated by \code{detect_coordinated_groups}}
}
\value{
a data.table with summary statistics for each user
}
\description{
Calculate user statistics: total posts shared, average time delta.
}
\details{
With this helper function you get a summary of the users, who
share coordinated content.
High number of posts shared and low average time delta are indicators for
highly coordinated (potentially automated) user behaviour.
}
