% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_stats.R
\name{user_stats}
\alias{user_stats}
\title{user_stats}
\usage{
user_stats(coord_graph, weight_threshold = c("full", "fast", "none"))
}
\arguments{
\item{weight_threshold}{The threshold to be used for filtering the graph (options: "full", "fast", or "none").}

\item{x}{a result data.table generated by \link{detect_groups}}
}
\value{
a data.table with summary statistics for each user
}
\description{
Calculate user statistics: total posts shared, average time delta, average edge symmetry score.
}
\details{
With this helper function, you can obtain summary statistics for the users in
the network. When applied to a network for which a narrower \code{time_window} has
been calculated using the \code{flag_speed_share} function, the summary statistics
are computed separately for both the full and faster networks. It is possible
to further refine the analysis by considering only the nodes that exceed the
edge weight threshold (reccomended). This filter can be applied to the full
network when no faster network is identified by the \code{flag_speed_share} function,
or when the network to be filtered is the full network. Alternatively, it can
be applied to the fast network when this is preliminarily identified via the
\code{flag_speed_share} function.
}
